refactor: move Milvus tests to their own module; extract _validate_document
